+++ reversi-server/controller/src/main/java/io/deeplay/camp/handlers/commands/MoveCommandHandler.java
@@ -4,8 +4,8 @@ import io.deeplay.camp.dto.BoardDTO;
 import io.deeplay.camp.entity.GameSession;
 import io.deeplay.camp.enums.GameStatus;
 import io.deeplay.camp.board.BoardService;
-import io.deeplay.camp.bot.BotStrategy;
-import io.deeplay.camp.bot.RandomBot;
+import io.deeplay.camp.bot.MyBots.BotStrategy;
+import io.deeplay.camp.bot.MyBots.RandomBot;
 import io.deeplay.camp.game.GameService;
 import io.deeplay.camp.handlers.main.MainHandler;
 import io.deeplay.camp.managers.SessionManager;
