+++ server.properties
@@ -1,4 +1,4 @@
 server.ip=127.0.0.1
 server.port=8080
 server.self-play=true
-server.game-count=10000
+server.game-count=100
